Added EText

+++ PnCEngine/Entity.h
@@ -8,7 +8,7 @@ struct Entity {
 	std::string src;
 	bool instantiated = false;
 
-	sf::Vector2i pos;
+	sf::Vector2f pos;
 	sf::Vector2f scale;
 	sf::Vector2f ori;
 	float rot;
